Introduced global gsap timelines
In this update, we've added a new feature to manage GSAP animations more efficiently. We've introduced a global object to store GSAP timelines and a function to create new ones if they don't exist. This change has been applied in two pages where the local GSAP timeline was replaced with the one from the global object.